[WNMGDS-980] - Convert UsaBanner component to typescript (#1377)
* UsaBanner ts conversion

* forgot exports

* forgot exports
