+++ packages/design-system-docs/src/pages/components/UsaBanner/_UsaBanner.docs.scss
@@ -17,7 +17,7 @@ Style guide: components.usa-banner
 
 @react-example UsaBanner.example.jsx
 
-@react-props UsaBanner.jsx
+@react-props UsaBanner.tsx
 
 Style guide: components.usa-banner.react
 */
