+++ packages/design-system/src/types/index.d.ts
@@ -12,7 +12,6 @@ export { default as Tabs } from './Tabs/Tabs';
 export { default as TabPanel } from './Tabs/TabPanel';
 export { default as Tooltip } from './Tooltip/Tooltip';
 export { default as TooltipIcon } from './Tooltip/TooltipIcon';
-export { default as UsaBanner } from './UsaBanner/UsaBanner';
 export { default as VerticalNav } from './VerticalNav/VerticalNav';
 
 // Type definitions generated by cmsds scripts
@@ -34,6 +33,7 @@ export * from './Table';
 export * from './TextField';
 export * from './Tooltip';
 export * from './Icons';
+export * from './UsaBanner';
 
 export * from './analytics';
 export * from './flags';
