+++ tsconfig.json
@@ -3,6 +3,7 @@
     "target": "es2015",
     "jsx": "react",
     "moduleResolution": "node",
+    "resolveJsonModule": true,
     "allowSyntheticDefaultImports": true,
     "esModuleInterop": true,
     "baseUrl": ".",
